+++ src/test/java/com/vonage/client/voice/StreamPayloadTest.java
@@ -15,22 +15,28 @@
  */
 package com.vonage.client.voice;
 
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertArrayEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import org.junit.jupiter.api.*;
+import java.net.URI;
 
 public class StreamPayloadTest {
     private StreamPayload payload;
 
     @BeforeEach
     public void setUp() throws Exception {
-        payload = new StreamPayload(
-                "https://nexmo-community.github.io/ncco-examples/assets/voice_api_audio_streaming.mp3",
-                2, 0.4, "8de6af25-5ee4-49ed-8283-de7ed108fb31");
+        payload = StreamPayload.builder()
+                .streamUrl("https://nexmo-community.github.io/ncco-examples/assets/voice_api_audio_streaming.mp3")
+                .loop(2).level(-0.4).build();
+        payload.setUuid("8de6af25-5ee4-49ed-8283-de7ed108fb31");
     }
 
     @Test
     public void getStreamUrl() throws Exception {
-        assertArrayEquals(new String[]{"https://nexmo-community.github.io/ncco-examples/assets/voice_api_audio_streaming.mp3"}, payload.getStreamUrl());
+        assertArrayEquals(new URI[]{URI.create(
+                "https://nexmo-community.github.io/ncco-examples/assets/voice_api_audio_streaming.mp3")},
+                payload.getStreamUrl()
+        );
     }
 
     @Test
@@ -40,14 +46,14 @@ public class StreamPayloadTest {
 
     @Test
     public void getLevel() throws Exception {
-        assertEquals(0.4, payload.getLevel(), 0.01);
+        assertEquals(-0.4, payload.getLevel(), 0.01);
     }
 
     @Test
     public void toJson() throws Exception {
         String jsonString = "{\"stream_url\":[" +
                 "\"https://nexmo-community.github.io/ncco-examples/assets/voice_api_audio_streaming.mp3\"" +
-                "],\"loop\":2,\"level\":0.4}";
+                "],\"loop\":2,\"level\":-0.4}";
         assertEquals(jsonString, payload.toJson());
     }